move _expand
